requirements.txt
beautifulsoup4==4.12.3
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
authlib==1.2.1
requests==2.31.0